requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
openai
python-dotenv
